feature(web-application) : move application-web schematics to applications/web
