--- src/lib/application-web/files/js/.babelrc
@@ -1,4 +1,0 @@
-{
-  "presets": ["@babel/preset-env"],
-  "plugins": [["@babel/plugin-proposal-decorators", { "legacy": true }]]
-}
