--- src/lib/application-web/files/js/nest-cli.json
@@ -1,5 +1,0 @@
-{
-  "language": "js",
-  "collection": "@nestjs/schematics",
-  "sourceRoot": "src"
-}
