--- src/lib/application-web/files/js/package.json
@@ -1,43 +1,0 @@
-{
-  "name": "<%= name %>",
-  "version": "<%= version %>",
-  "description": "<%= description %>",
-  "author": "<%= author %>",
-  "license": "MIT",
-  "scripts": {
-    "format": "prettier --write \"**/*.js\"",
-    "start": "babel-node index.js",
-    "start:dev": "nodemon",
-    "test": "jest",
-    "test:cov": "jest --coverage",
-    "test:e2e": "jest --config ./test/jest-e2e.json"
-  },
-  "dependencies": {
-    "@nestjs/common": "^6.0.0",
-    "@nestjs/core": "^6.0.0", 
-    "@nestjs/platform-express": "^6.0.0",
-    "reflect-metadata": "^0.1.12",
-    "rxjs": "^6.3.3"
-  },
-  "devDependencies": {
-    "@nestjs/testing": "^6.0.0",
-    "@babel/core": "^7.4.3",
-    "@babel/node": "^7.2.2",
-    "@babel/plugin-proposal-decorators": "^7.4.0",
-    "@babel/preset-env": "^7.4.3",
-    "@babel/register": "^7.4.0",
-    "jest": "^22.4.3",
-    "nodemon": "^1.14.1",
-    "prettier": "^1.11.1",
-    "supertest": "^3.0.0"
-  },
-  "jest": {
-    "moduleFileExtensions": [
-      "js",
-      "json"
-    ],
-    "rootDir": "src",
-    "testRegex": ".spec.js$",
-    "coverageDirectory": "../coverage"
-  }
-}
