--- src/lib/application-web/files/ts/.prettierrc
@@ -1,4 +1,0 @@
-{
-  "singleQuote": true,
-  "trailingComma": "all"
-}
