--- src/lib/application-web/files/ts/nest-cli.json
@@ -1,5 +1,0 @@
-{
-  "language": "ts",
-  "collection": "@nestjs/schematics",
-  "sourceRoot": "src"
-}
