+++ src/lib/applications/web/files/js/test/app.e2e-spec.js
