+++ src/lib/applications/web/schema.json
