--- src/lib/project/constantes.ts
@@ -1,9 +1,0 @@
-export enum DefaultOption {
-  AUTHOR = '',
-  DESCRIPTION = '',
-  LANGUAGE = 'ts',
-  VERSION = '0.0.1',
-  PATH_NAME = 'src',
-  LIB_PATH = 'libs',
-  PACKAGE_MANAGER = 'npm',
-}
